라이브러리 그룹폴더 삭제기능 구현

+++ app/src/main/java/com/cow/bridge/library/activity/GroupDetailActivity.kt
@@ -1,8 +1,10 @@
 package com.cow.bridge.library.activity
 
+import android.content.DialogInterface
 import android.content.Intent
 import android.content.SharedPreferences
 import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.LinearLayoutManager
@@ -10,6 +12,8 @@ import android.util.Log
 import com.bumptech.glide.Glide
 import com.cow.bridge.R
 import com.cow.bridge.library.adapter.GroupContentsAdapter
+import com.cow.bridge.library.dialog.LibraryDeleteDialog
+import com.cow.bridge.library.dialog.LibraryDialog
 import com.cow.bridge.model.Group
 import com.cow.bridge.network.ApplicationController
 import com.cow.bridge.network.Network
@@ -69,23 +73,64 @@ class GroupDetailActivity : AppCompatActivity() {
         })
 
         group_text_delete.setOnClickListener {
-            var messagesCall = api?.deleteGroup(Group(group.groupIdx))
-            messagesCall?.enqueue(object : Callback<Network> {
-                override fun onResponse(call: Call<Network>?, response: Response<Network>?) {
-                    var network = response!!.body()
-                    Log.v("test", Gson().toJson(network))
-                    if(network?.message.equals("ok")){
-                        finish()
+            val libraryDeleteDialog : LibraryDeleteDialog = LibraryDeleteDialog(this@GroupDetailActivity)
+            libraryDeleteDialog.window.setBackgroundDrawable(ColorDrawable(android.graphics.Color.TRANSPARENT))
+            libraryDeleteDialog.show()
+            libraryDeleteDialog.setOnDismissListener(object : DialogInterface.OnDismissListener{
+                override fun onDismiss(dialog: DialogInterface?) {
+                    with(dialog as LibraryDeleteDialog){
+                        if(confirm){
+                            var messagesCall = api?.deleteGroup(Group(group.groupIdx))
+                            messagesCall?.enqueue(object : Callback<Network> {
+                                override fun onResponse(call: Call<Network>?, response: Response<Network>?) {
+                                    var network = response!!.body()
+                                    Log.v("test", Gson().toJson(network))
+                                    if(network?.message.equals("ok")){
+                                        finish()
+                                    }
+                                }
+                                override fun onFailure(call: Call<Network>?, t: Throwable?) {
+
+                                }
+                            })
+                        }
                     }
                 }
-                override fun onFailure(call: Call<Network>?, t: Throwable?) {
 
-                }
             })
+
         }
 
         group_text_edit.setOnClickListener {
+            val libraryDialog : LibraryDialog = LibraryDialog(this@GroupDetailActivity, "#FFFFFF", group.groupTitle)
+            libraryDialog.window.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+            libraryDialog.show()
+            libraryDialog.setOnDismissListener {dialog ->
+                with(dialog as LibraryDialog){
+                    if(confirm){
+                        var sp : SharedPreferences = getSharedPreferences("bridge", AppCompatActivity.MODE_PRIVATE)
+                        var messagesCall = api?.modifyGroup(Group(1, group.groupIdx, groupName!!, groupColor!!))
+                        messagesCall?.enqueue(object : Callback<Network> {
+                            override fun onResponse(call: Call<Network>?, response: Response<Network>?) {
+                                var network = response!!.body()
+                                Log.v("test", Gson().toJson(network))
+                                if(network?.message.equals("ok")){
+                                    this@GroupDetailActivity.group_text_name.text = groupName
+                                    try{
+                                        this@GroupDetailActivity.group_image_bg.setBackgroundColor(Color.parseColor(groupColor))
+                                    }catch (e : IllegalArgumentException){
+                                        //this@GroupDetailActivity.group_image_bg.setBackgroundColor(Color.parseColor("#F7F7F7"))
+                                    }
+                                }
+                            }
+                            override fun onFailure(call: Call<Network>?, t: Throwable?) {
+                                Log.v("test", t.toString())
 
+                            }
+                        })
+                    }
+                }
+            }
         }
 
         group_image_back.setOnClickListener { finish() }
